Implemented Review comments
